pyproject.toml
[project]
name = "ontology-engine"
version = "0.1.0"
description = "Structured knowledge extraction from meeting transcripts → PostgreSQL ontology graph"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Yann Guo", email = "yann@whitemirror.ai" },
]
keywords = ["ontology", "knowledge-graph", "meeting", "extraction", "llm", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "asyncpg>=0.29.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
gemini = ["google-genai>=1.0.0"]
openai = ["openai>=1.30.0"]
all = ["google-genai>=1.0.0", "openai>=1.30.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
ontology-engine = "ontology_engine.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ontology_engine"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
